LPS-174140 Fix One
Prevent the portal from creating _release entries here.
This is similiar to the initial solution of using the database auto upgrade.  If the portal is not the latest schema version, it will not create the release_ entries.
I did have to use DataAccess becuase the global variable connection is null.