LPS-174140 Fix Two
Prevent the portal from launching if they are not on the latest schema
version.  This does not prevent the user from using the schema version on
different portal versions that are close (e.g u63->u64).  However I don't think the issue would be reproduce in the cases that that schema version is the same.