LPS-185647 Perhaps a bit confusingly, but fortunately for us, the FlushEventEntityListener is responsible for actually executing the flush logic. This means that we can simply wrap the flush logic inside of a CompanyThreadLocal.setInitializingCompanyIdWithSafeCloseable when it is appropriate to do so (i.e. when the model is sharded).
Important note: The reason that we use
setInitializingCompanyIdWithSafeCloseable here instead of the standard
setWithSafeCloseable is to prevent a potential infinite loop or
StackOverflowError. For example, if we are flushing a User, then the
following could occur if we used setWithSafeCloseable:

1. Detect that we need to change the companyId before flushing.
2. Call setCompanyIdWithSafeCloseable.
3. When setting the companyId, it sets the UserThreadLocals.
4. In order to do so, it tries to fetch the default User.
5. When fetching the default User, it detects that the session is dirty
   and so it needs to flush first.
6. Return to step 1 and loop until a StackOverflowError occurs.

Calling setInitializingCompanyIdWithSafeCloseable does not initialize
the UserThreadLocals or do anything other than set the ThreadLocal for
companyId, so it avoids step 3 and breaks the possible infinite loop.

And we don't need any of that UserThreadLocal stuff for this case anyway
since everything inside of here is just third-party Hibernate code;
there should be no Liferay code that would be dependent on any
UserThreadLocal.